requirements.txt
ipykernel==6.29.5
pandas==2.2.3
plotly==5.24.1
python-binance==1.0.27
winloop==0.1.8
asyncio==3.4.3